[Darwin] Move Dnssd related files under dnssd/ and rename MdnsError => DnssdError (#39438)
