+++ src/platform/Darwin/BUILD.gn
@@ -57,14 +57,7 @@ static_library("Darwin") {
     "ConnectivityManagerImpl.h",
     "DiagnosticDataProviderImpl.cpp",
     "DiagnosticDataProviderImpl.h",
-    "DnssdContexts.cpp",
-    "DnssdHostNameRegistrar.mm",
-    "DnssdImpl.cpp",
-    "DnssdImpl.h",
-    "DnssdType.cpp",
     "InetPlatformConfig.h",
-    "MdnsError.cpp",
-    "MdnsError.h",
     "NetworkCommissioningDriver.h",
     "PlatformManagerImpl.h",
     "PlatformManagerImpl.mm",
@@ -75,6 +68,13 @@ static_library("Darwin") {
     "SystemTimeSupport.cpp",
     "UserDefaults.h",
     "UserDefaults.mm",
+    "dnssd/DnssdContexts.cpp",
+    "dnssd/DnssdError.cpp",
+    "dnssd/DnssdError.h",
+    "dnssd/DnssdHostNameRegistrar.mm",
+    "dnssd/DnssdImpl.cpp",
+    "dnssd/DnssdImpl.h",
+    "dnssd/DnssdType.cpp",
     "inet/NetworkMonitor.h",
     "inet/NetworkMonitor.mm",
     "system/SystemLayerImplDispatch.h",
