+++ src/platform/Darwin/dnssd/DnssdError.cpp
