+++ src/platform/Darwin/dnssd/DnssdError.h
